requirements.txt
Django==2.0.9
Pillow==5.3.0
autopep8==1.4.2
dj-database-url==0.5.0
django-allauth==0.38.0
djangorestframework==3.9.0
gnureadline==6.3.8
gunicorn==19.9.0
psycopg2==2.7.5
requests==2.20.0
rollbar==0.13.13
six==1.11.0
stream-django==1.4.0
whitenoise==4.1
